ONG251-24 refactor UserMapper
