+++ src/main/java/com/alkemy/ong/dto/response/user/UserResponse.java
@@ -1,12 +1,15 @@
 package com.alkemy.ong.dto.response.user;
 
+import java.util.UUID;
+
 import lombok.Data;
 
 @Data
 public class UserResponse {
-		private String firstName;
-		private String lastName;
-		private String email;
-		private String photo;
-		private String role;
+	private UUID id;
+	private String firstName;
+	private String lastName;
+	private String email;
+	private String photo;
+	private String role;
 }
