+++ src/main/java/com/alkemy/ong/mappers/UserMapper.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.config.mapper;
+package com.alkemy.ong.mappers;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -16,6 +16,7 @@ public class UserMapper {
 
 	public UserResponse mapResponse(User user) {
 		UserResponse userResponse = new UserResponse();
+		userResponse.setId(user.getId());
 		userResponse.setFirstName(user.getFirstName());
 		userResponse.setLastName(user.getLastName());
 		userResponse.setEmail(user.getEmail());
