+++ src/main/java/com/alkemy/ong/services/imp/UserServiceImp.java
@@ -3,10 +3,10 @@ package com.alkemy.ong.services.imp;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import com.alkemy.ong.config.mapper.UserMapper;
 import com.alkemy.ong.dto.request.user.UserRegisterRequest;
 import com.alkemy.ong.dto.response.user.UserResponse;
 import com.alkemy.ong.exception.EmailAlreadyExistsException;
+import com.alkemy.ong.mappers.UserMapper;
 import com.alkemy.ong.models.User;
 import com.alkemy.ong.repositories.UserRepository;
 import com.alkemy.ong.services.IUserService;
